feat: correspond Kubernetes resource pools to namespaces
This allows each resource pool in a Kubernetes cluster to map to a
specific namespace in the cluster. The pods actor now maintains pod and
config map clients for each namespace that might be necessary and passes
the appropriate one along to each of its child actors. (Since the
clients have namespaces built into them that can't be changed, we can't
just have one of each kind of client.)

An alternative would be to have a separate pods actor for each
namespace, along with the attending actors for informers, listeners, and
the request queue. That has a certain elegance and might be a smaller
code change, but I didn't go that way because it seemed kind of
heavyweight in execution.